backend/requirements.txt
fastapi
sqlalchemy
passlib[bcrypt]
pydantic[dotenv]
python-multipart
PyJWT
fastapi_mail
jinja2
uvicorn
requests